@@ -1,36 +1,41 @@
 {
   "name": "astraeus-website",
   "version": "0.1.0",
+  "type": "module",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "vite dev",
+    "build": "vite build",
+    "lint": "vite lint"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.15",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@types/node": "20.4.4",
-    "@types/react": "18.2.15",
-    "@types/react-dom": "18.2.7",
-    "ahooks": "^3.7.8",
+    "@solidjs/router": "^0.14.1",
     "autoprefixer": "10.4.14",
-    "eslint": "8.45.0",
-    "eslint-config-next": "13.4.12",
     "lodash": "^4.17.21",
-    "next": "13.4.12",
     "postcss": "8.4.27",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-markdown": "^9.0.1",
-    "react-redux": "^8.1.1",
-    "redux": "^4.2.1",
-    "tailwindcss": "3.3.3",
-    "typescript": "5.1.6"
+    "solid-icons": "1.0.12",
+    "solid-js": "^1.8.18",
+    "solidjs-use": "^2.3.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.195",
-    "sass": "^1.64.1"
+    "@types/lodash": "^4.17.6",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-import-resolver-alias": "^1.1.2",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-json": "^3.1.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-solid": "^0.13.1",
+    "eslint-plugin-testing-library": "^4.12.4",
+    "prettier": "^3.2.4",
+    "sass": "^1.77.8",
+    "tailwindcss": "3.4.4",
+    "typescript": "5.5.3",
+    "vite": "^5.0.0",
+    "vite-plugin-pages": "^0.32.1",
+    "vite-plugin-solid": "^2.8.0"
   }
 }
